Can send an auth request and message when the form is submitted, but it doesn't really work work.
I need to have a client that is running constantly and updating a central record of statuses, and the web app needs to interact with the same model. So the web app pushes an 'auth me' type request onto the queue, the client pops jobs off the queue and handles all auth request, communication, etc and updates the queue (or adds a new job, or whatever).

The web app needs to be polling the queue constantly (through a heartbeat or open comet-style connection from the browser probably?) and then updates when it sees the results of the process.

There is a very rough client in play, based on some example code that handles messages coming back, but the job queue doesn't exist. There's a fairly standard queue-style system for RoR called Delayed Jobs that I should probably have the two sides work from.

A trick for me will be to figure out how to get the client, which is not a Rails app, talking to the same model that the Rails app does - or how to make it a command-line based rails app which will give me the ActiveRecord stuff for free.